refactor: split PriceAlert page into modular components
